+++ example/urls_test.py
@@ -44,30 +44,30 @@ urlpatterns = [
     url(r'^', include(router.urls)),
 
     # old tests
-    url(r'identities/default/(?P<pk>\d+)',
+    url(r'identities/default/(?P<pk>\d+)$',
         GenericIdentity.as_view(), name='user-default'),
 
 
-    url(r'^entries/(?P<entry_pk>[^/.]+)/blog',
+    url(r'^entries/(?P<entry_pk>[^/.]+)/blog$',
         BlogViewSet.as_view({'get': 'retrieve'}),
         name='entry-blog'
         ),
-    url(r'^entries/(?P<entry_pk>[^/.]+)/comments',
+    url(r'^entries/(?P<entry_pk>[^/.]+)/comments$',
         CommentViewSet.as_view({'get': 'list'}),
         name='entry-comments'
         ),
-    url(r'^entries/(?P<entry_pk>[^/.]+)/suggested/',
+    url(r'^entries/(?P<entry_pk>[^/.]+)/suggested/$',
         EntryViewSet.as_view({'get': 'list'}),
         name='entry-suggested'
         ),
-    url(r'^drf-entries/(?P<entry_pk>[^/.]+)/suggested/',
+    url(r'^drf-entries/(?P<entry_pk>[^/.]+)/suggested/$',
         DRFEntryViewSet.as_view({'get': 'list'}),
         name='drf-entry-suggested'
         ),
-    url(r'entries/(?P<entry_pk>[^/.]+)/authors',
+    url(r'entries/(?P<entry_pk>[^/.]+)/authors$',
         AuthorViewSet.as_view({'get': 'list'}),
         name='entry-authors'),
-    url(r'entries/(?P<entry_pk>[^/.]+)/featured',
+    url(r'entries/(?P<entry_pk>[^/.]+)/featured$',
         EntryViewSet.as_view({'get': 'retrieve'}),
         name='entry-featured'),
 
@@ -75,16 +75,16 @@ urlpatterns = [
         AuthorViewSet.as_view({'get': 'retrieve_related'}),
         name='author-related'),
 
-    url(r'^entries/(?P<pk>[^/.]+)/relationships/(?P<related_field>\w+)',
+    url(r'^entries/(?P<pk>[^/.]+)/relationships/(?P<related_field>\w+)$',
         EntryRelationshipView.as_view(),
         name='entry-relationships'),
-    url(r'^blogs/(?P<pk>[^/.]+)/relationships/(?P<related_field>\w+)',
+    url(r'^blogs/(?P<pk>[^/.]+)/relationships/(?P<related_field>\w+)$',
         BlogRelationshipView.as_view(),
         name='blog-relationships'),
-    url(r'^comments/(?P<pk>[^/.]+)/relationships/(?P<related_field>\w+)',
+    url(r'^comments/(?P<pk>[^/.]+)/relationships/(?P<related_field>\w+)$',
         CommentRelationshipView.as_view(),
         name='comment-relationships'),
-    url(r'^authors/(?P<pk>[^/.]+)/relationships/(?P<related_field>\w+)',
+    url(r'^authors/(?P<pk>[^/.]+)/relationships/(?P<related_field>\w+)$',
         AuthorRelationshipView.as_view(),
         name='author-relationships'),
 ]